施工回退

+++ app/Http/Controllers/OrderQuotationController.php
@@ -462,7 +462,7 @@ class OrderQuotationController extends Controller
                     ->where('company_id', $order->wusun_company_id)
                     ->first();
                 if ($approvalOrder) ApprovalOrderProcess::where('approval_order_id', $approvalOrder->id)->delete();
-                $approvalOrder->delete();
+                if ($approvalOrder) $approvalOrder->delete();
 
                 $approvalOrder = ApprovalOrder::create([
                     'order_id' => $order->id,
